requirements.txt
Flask
Flask-JWT-Extended
Flask-Bcrypt
Flask-Cors
pymongo
python-dotenv
